Cache Counterpoint damage units and split action data
